pyproject.toml
[project]
name = "omniemployee"
version = "0.1.0"
description = "A powerful and extensible base agent system with MCP and Skills support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
